cleaner refactor: Don't flush empty containers
It's a fix for a time-of-check-time-of-use problem during populating flush
containers in _ocf_mngt_get_flush_container().

Although the flush containers are populated under the global metadata
write lock it is still possible that concurrent IOs will mark some cache
lines as clean between the flushing thread reads core->dirty_clines and
tests status bits of all cache lines. If the cleaned cache lines happened
to be the last dirty cache lines for that core, we end up with an empty
flush container.

Signed-off-by: Michal Mielewczyk <michal.mielewczyk@huawei.com>