cleaner: Update progress before submiting request
The flush is an async operation so there is a slight chance that
the completion would be called before the progress is updated. Since
the completion uses this information to determine whether to continue
flushing it could lead to a race condition and e.g. submitting an empty
flush request

Signed-off-by: Michal Mielewczyk <michal.mielewczyk@huawei.com>