Sort flush data only if there is anything to flush
If there is nothing to flush, the env_sort() is called with NULL pointer
instead of the first element of array to sort.
This could be fixed in specific implementation of env_sort() to return
immediately if number of elements is 0 but all other routines check
the element count before calling ocf_cleaner_sort_flush_data() except
of ocf_cleaner_sort_flush_containers(), so putting this check there seems
more appropriate and makes flush data sorting less prone to errors.
Moreover, improve it even a bit further by calling the sorting routine
only if there is more then one element.

Signed-off-by: Rafal Stefanowski <rafal.stefanowski@huawei.com>